# server/gen/swagger/swagger.yaml
basePath: /api/v1
definitions:
  apierror.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  sqlc.EntityType:
    enum:
    - story
    - chapter
    - comment
    - submission
    type: string
    x-enum-varnames:
    - EntityTypeStory
    - EntityTypeChapter
    - EntityTypeComment
    - EntityTypeSubmission
...
  tag.CreateTagRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 255
        type: string
      entity_id:
        type: string
      entity_type:
        allOf:
        - $ref: '#/definitions/sqlc.EntityType'
        enum:
        - story
        - chapter
        - comment
        - submission
      is_hidden:
        type: boolean
      is_recommended:
        type: boolean
      is_system:
        type: boolean
      name:
        maxLength: 50
        type: string
    required:
    - color
    - entity_type
    - name
    type: object
  tag.TagListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      tags:
        items:
          $ref: '#/definitions/tag.TagResponse'
        type: array
      total:
        type: integer
    type: object
  tag.TagResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      entity_id:
        type: string
      entity_type:
        $ref: '#/definitions/sqlc.EntityType'
      id:
        type: string
      is_hidden:
        type: boolean
      is_recommended:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
  tag.UpdateTagRequest:
    properties:
      color:
        type: string
      description:
        maxLength: 255
        type: string
      is_hidden:
        type: boolean
      is_recommended:
        type: boolean
      is_system:
        type: boolean
      name:
        maxLength: 50
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: trao0312@gmail.com
    name: API Support
    url: https://github.com/justblue/samsa
  description: Samsa is a simple backend for a writing platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Samsa
  version: 0.1.0
paths:
...
  /entities/{entity_type}/{entity_id}/tags:
    get:
      description: Retrieves all tags for a specific entity. Requires `user` actor
        and `tag:read` scope.
      parameters:
      - description: Entity type (story, chapter, comment, submission)
        in: path
        name: entity_type
        required: true
        type: string
      - description: Entity UUID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.TagResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get entity tags
      tags:
      - tags
  /entities/{entity_type}/{entity_id}/tags/count:
    get:
      description: Retrieves the count of tags for a specific entity. Requires `user`
        actor and `tag:read` scope.
      parameters:
      - description: Entity type
        in: path
        name: entity_type
        required: true
        type: string
      - description: Entity UUID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get tag count
      tags:
      - tags
...
  /owners/{owner_id}/tags:
    get:
      description: Retrieves all tags owned by a specific user. Requires `user` actor
        and `tag:read` scope.
      parameters:
      - description: Owner UUID
        in: path
        name: owner_id
        required: true
        type: string
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get owner tags
      tags:
      - tags
...
  /tags:
    get:
      description: Retrieves a paginated list of tags with optional filters. Requires
        `user` actor and `tag:read` scope.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by owner UUID
        in: query
        name: owner_id
        type: string
      - description: Filter by entity UUID
        in: query
        name: entity_id
        type: string
      - description: Filter by entity type
        in: query
        name: entity_type
        type: string
      - description: Filter by hidden status
        in: query
        name: is_hidden
        type: boolean
      - description: Filter by system tags
        in: query
        name: is_system
        type: boolean
      - description: Filter by recommended tags
        in: query
        name: is_recommended
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: List tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Creates a new tag. Requires `user` actor and `tag:write` scope.
      parameters:
      - description: Tag creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierror.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - tags
  /tags/{tag_id}:
    delete:
      description: Deletes a tag. Only the tag owner can delete. Requires `user` actor
        and `tag:write` scope.
      parameters:
      - description: Tag UUID
        in: path
        name: tag_id
        required: true
        type: string
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - tags
    get:
      description: Retrieves a tag by its UUID. Requires `user` actor and `tag:read`
        scope.
      parameters:
      - description: Tag UUID
        in: path
        name: tag_id
        required: true
        type: string
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get tag by ID
      tags:
      - tags
    patch:
      consumes:
      - application/json
      description: Updates a tag. Only the tag owner can update. Requires `user` actor
        and `tag:write` scope.
      parameters:
      - description: Tag UUID
        in: path
        name: tag_id
        required: true
        type: string
      - description: Tag update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tag.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - tags
  /tags/batch:
    get:
      description: Retrieves tags by their UUIDs. Requires `user` actor and `tag:read`
        scope.
      parameters:
      - description: Comma-separated tag UUIDs
        in: query
        name: ids
        required: true
        type: string
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.TagResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get tags by IDs
      tags:
      - tags
  /tags/search:
    get:
      description: Searches for tags by name. Requires `user` actor and `tag:read`
        scope.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Search tags
      tags:
      - tags
...
swagger: "2.0"